+++ src/commands/up/index.ts
@@ -4,7 +4,7 @@ import { S3 } from '@aws-sdk/client-s3';
 import { ChildProcess } from 'child_process';
 import isNil from 'lodash.isnil';
 import logger from '../../logger';
-import { isPortAvailable, logAndThrow, editDockerFile, runCommand, runCommandSync, streamToFile } from '../../utils/os';
+import { isPortAvailable, logAndThrow, editDockerFile, runCommand, runCommandSync, streamToFile, validateCwdPermissions } from '../../utils/os';
 import { ImageArchitecture, UpOptions } from '../../types';
 import { DEFAULT_CONFIG_FILENAME, Platform, SEP } from '../../constants';
 import { parseConfig, validateDependencies } from '../../utils/config';
@@ -70,6 +70,8 @@ async function getDependencies (file: string, parentDirectory: string) {
 }
 
 async function pullDockerFiles (tag: string) {
+  validateCwdPermissions();
+
   const s3Client = new S3({
     signer: { sign: async (request: any) => request },
     region: 'us-west-2'
@@ -78,13 +80,15 @@ async function pullDockerFiles (tag: string) {
     Bucket: 'ops-stacks-config-storage-bucket-us-west-2',
     Key: 'Dockerfile.api'
   });
-  await streamToFile(apiRes.Body, Platform.ApiFilePath);
-  editDockerFile(Platform.ApiFilePath, tag);
   const uiRes = await s3Client.getObject({
     Bucket: 'ops-stacks-config-storage-bucket-us-west-2',
     Key: 'Dockerfile.ui'
   });
+
+  await streamToFile(apiRes.Body, Platform.ApiFilePath);
   await streamToFile(uiRes.Body, Platform.UiFilePath);
+  
+  editDockerFile(Platform.ApiFilePath, tag);
   editDockerFile(Platform.UiFilePath, tag);
 }